refactor: 가독성 향상을 위한 DTO 패키지 리팩토링
