+++ src/main/java/com/project/baedalsodae/store/controller/AdminStoreController.java
@@ -3,8 +3,8 @@ package com.project.baedalsodae.store.controller;
 import com.project.baedalsodae.auth.security.UserDetailsImpl;
 import com.project.baedalsodae.global.common.ApiResponse;
 import com.project.baedalsodae.global.common.SuccessCode;
-import com.project.baedalsodae.store.dto.request.UpdateStoreRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreStatusRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreStatusRequest;
 import com.project.baedalsodae.store.service.AdminStoreService;
 import jakarta.validation.Valid;
 import java.util.UUID;
