+++ src/main/java/com/project/baedalsodae/store/controller/StoreController.java
@@ -3,14 +3,14 @@ package com.project.baedalsodae.store.controller;
 import com.project.baedalsodae.auth.security.UserDetailsImpl;
 import com.project.baedalsodae.global.common.ApiResponse;
 import com.project.baedalsodae.global.common.SuccessCode;
-import com.project.baedalsodae.store.dto.request.CreateStoreRequest;
-import com.project.baedalsodae.store.dto.request.StoreCursorRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreStatusRequest;
-import com.project.baedalsodae.store.dto.response.OwnerStoreResponse;
-import com.project.baedalsodae.store.dto.response.StoreDetailResponse;
-import com.project.baedalsodae.store.dto.response.StorePageResponse;
-import com.project.baedalsodae.store.dto.response.StoreSearchPageResponse;
+import com.project.baedalsodae.store.dto.request.store.CreateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.StoreCursorRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreStatusRequest;
+import com.project.baedalsodae.store.dto.response.store.OwnerStoreResponse;
+import com.project.baedalsodae.store.dto.response.store.StoreDetailResponse;
+import com.project.baedalsodae.store.dto.response.store.StorePageResponse;
+import com.project.baedalsodae.store.dto.response.store.StoreSearchPageResponse;
 import com.project.baedalsodae.store.entity.enums.SortType;
 import com.project.baedalsodae.store.service.StoreCommandService;
 import com.project.baedalsodae.store.service.StoreQueryService;
