+++ src/main/java/com/project/baedalsodae/store/controller/StoreHourController.java
@@ -3,8 +3,8 @@ package com.project.baedalsodae.store.controller;
 import com.project.baedalsodae.auth.security.UserDetailsImpl;
 import com.project.baedalsodae.global.common.ApiResponse;
 import com.project.baedalsodae.global.common.SuccessCode;
-import com.project.baedalsodae.store.dto.request.StoreHoursRequest;
-import com.project.baedalsodae.store.dto.response.StoreHoursResponse;
+import com.project.baedalsodae.store.dto.request.store.StoreHoursRequest;
+import com.project.baedalsodae.store.dto.response.store.StoreHoursResponse;
 import com.project.baedalsodae.store.service.StoreHoursService;
 import jakarta.validation.Valid;
 import java.util.List;
