+++ src/main/java/com/project/baedalsodae/store/dto/request/store/StoreCursorRequest.java
@@ -1,4 +1,4 @@
-package com.project.baedalsodae.store.dto.request;
+package com.project.baedalsodae.store.dto.request.store;
 
 import com.project.baedalsodae.store.entity.enums.SortType;
 import jakarta.validation.constraints.Min;
