+++ src/main/java/com/project/baedalsodae/store/dto/response/store/OwnerStoreResponse.java
@@ -1,4 +1,4 @@
-package com.project.baedalsodae.store.dto.response;
+package com.project.baedalsodae.store.dto.response.store;
 
 import com.project.baedalsodae.global.common.dto.AddressResponse;
 import com.project.baedalsodae.store.entity.Store;
