+++ src/main/java/com/project/baedalsodae/store/repository/custom/StoreCustomRepository.java
@@ -1,6 +1,6 @@
 package com.project.baedalsodae.store.repository.custom;
 
-import com.project.baedalsodae.store.dto.request.StoreCursorRequest;
+import com.project.baedalsodae.store.dto.request.store.StoreCursorRequest;
 import com.project.baedalsodae.store.entity.Store;
 import com.project.baedalsodae.store.entity.enums.SortType;
 import java.util.List;
