+++ src/main/java/com/project/baedalsodae/store/service/StoreCommandService.java
@@ -1,7 +1,7 @@
 package com.project.baedalsodae.store.service;
 
-import com.project.baedalsodae.store.dto.request.CreateStoreRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.CreateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreRequest;
 import com.project.baedalsodae.store.entity.enums.StoreStatus;
 import java.util.UUID;
