+++ src/main/java/com/project/baedalsodae/store/service/StoreHoursService.java
@@ -1,8 +1,8 @@
 package com.project.baedalsodae.store.service;
 
 import com.project.baedalsodae.auth.security.UserDetailsImpl;
-import com.project.baedalsodae.store.dto.request.StoreHoursRequest;
-import com.project.baedalsodae.store.dto.response.StoreHoursResponse;
+import com.project.baedalsodae.store.dto.request.store.StoreHoursRequest;
+import com.project.baedalsodae.store.dto.response.store.StoreHoursResponse;
 import java.util.List;
 import java.util.UUID;
