+++ src/main/java/com/project/baedalsodae/store/service/impl/StoreCommandServiceImpl.java
@@ -6,8 +6,8 @@ import static com.project.baedalsodae.store.entity.enums.StoreStatus.SUSPENDED;
 import com.project.baedalsodae.global.common.BusinessException;
 import com.project.baedalsodae.global.common.ErrorCode;
 import com.project.baedalsodae.global.common.entity.Address;
-import com.project.baedalsodae.store.dto.request.CreateStoreRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.CreateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreRequest;
 import com.project.baedalsodae.store.entity.Store;
 import com.project.baedalsodae.store.entity.StoreCategory;
 import com.project.baedalsodae.store.entity.enums.StoreStatus;
