+++ src/main/java/com/project/baedalsodae/store/service/impl/StoreHoursServiceImpl.java
@@ -4,8 +4,8 @@ import com.project.baedalsodae.auth.security.UserDetailsImpl;
 import com.project.baedalsodae.global.common.BusinessException;
 import com.project.baedalsodae.global.common.ErrorCode;
 import com.project.baedalsodae.menu.common.StoreOwnershipValidator;
-import com.project.baedalsodae.store.dto.request.StoreHoursRequest;
-import com.project.baedalsodae.store.dto.response.StoreHoursResponse;
+import com.project.baedalsodae.store.dto.request.store.StoreHoursRequest;
+import com.project.baedalsodae.store.dto.response.store.StoreHoursResponse;
 import com.project.baedalsodae.store.entity.Store;
 import com.project.baedalsodae.store.entity.StoreHours;
 import com.project.baedalsodae.store.repository.StoreHoursRepository;
