+++ src/test/java/com/project/baedalsodae/store/controller/StoreControllerTest.java
@@ -17,11 +17,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.baedalsodae.auth.security.UserDetailsImpl;
 import com.project.baedalsodae.global.common.dto.AddressRequest;
 import com.project.baedalsodae.global.common.dto.AddressResponse;
-import com.project.baedalsodae.store.dto.request.CreateStoreRequest;
-import com.project.baedalsodae.store.dto.request.StoreCursorRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreRequest;
-import com.project.baedalsodae.store.dto.request.UpdateStoreStatusRequest;
-import com.project.baedalsodae.store.dto.response.*;
+import com.project.baedalsodae.store.dto.request.store.CreateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.StoreCursorRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreRequest;
+import com.project.baedalsodae.store.dto.request.store.UpdateStoreStatusRequest;
+import com.project.baedalsodae.store.dto.response.store.*;
 import com.project.baedalsodae.store.entity.enums.SortType;
 import com.project.baedalsodae.store.entity.enums.StoreStatus;
 import com.project.baedalsodae.store.service.StoreCommandService;
